removed hybridRelationships, moved parsing funcs
