+++ js/components/graph/Graph.js
@@ -7,7 +7,6 @@ import Button from "./Button";
 import EdgeGroup from "./EdgeGroup";
 import InfoBox from "./InfoBox";
 import NodeGroup from "./NodeGroup";
-import {parseAnd} from "./NodeGroup";
 import RegionGroup from "./RegionGroup";
 import * as focusInfo from "./sidebar/focus_descriptions";
 
@@ -110,71 +109,6 @@ export default class Graph extends React.Component {
       .getElementById("react-graph")
       .removeEventListener("wheel", this.onWheel);
   }
-/** Adds parents of hybridNode to the parents object, under it's id, and populates
-   * hybridRelationships with
-   *
-   * @param {Node} hybridNode
-   * @param {Array} nodesJSON
-   * @param {Array} hybridRelationships
-   * @param {Object} parents
-   */
-  populateParentsAndHybridRelationship(hybridNode, nodesJSON, hybridRelationships, parents){
-
-    /**
-     * Helper for hybrid computation. Finds the node with the same course label as the hybrid.
-     * @param  {string} course
-     * @param {Array} nodesJSON
-     * @return {Node}
-     */
-    var findRelationship = (course, nodesJSON) => {
-      var nodes = nodesJSON;
-      var node = nodes.find(
-        n =>
-          n.type_ === "Node" &&
-          n.text.some(textTag => textTag.text.includes(course))
-      );
-      return node;
-    }
-    // parse prereqs based on text
-    var hybridText = "";
-    hybridNode.text.forEach(textTag => (hybridText += textTag.text));
-    var nodeParents = [];
-    // First search for entire string (see Stats graph)
-    var prereqNode = findRelationship(hybridText, nodesJSON);
-    if (prereqNode !== undefined) {
-      nodeParents.push(prereqNode.id_);
-      hybridRelationships.push([prereqNode.id_, hybridNode.id_]);
-    } else {
-      // Parse text first
-      var prereqs = parseAnd(hybridText)[0];
-      prereqs.forEach(course => {
-        if (typeof course === "string") {
-          prereqNode = findRelationship(course, nodesJSON);
-          if (prereqNode !== undefined) {
-            nodeParents.push(prereqNode.id_);
-            hybridRelationships.push([prereqNode.id_, hybridNode.id_]);
-          } else {
-            console.error("Could not find prereq for ", hybridText);
-          }
-        } else if (typeof course === "object") {
-          var orPrereq = [];
-          course.forEach(c => {
-            var prereqNode = findRelationship(c, nodesJSON);
-            if (prereqNode !== undefined) {
-              orPrereq.push(prereqNode.id_);
-              hybridRelationships.push([prereqNode.id_, hybridNode.id_]);
-            } else {
-              console.error("Could not find prereq for ", hybridText);
-            }
-          });
-          if (orPrereq.length > 0) {
-            nodeParents.push(orPrereq);
-          }
-        }
-      });
-    }
-  parents[hybridNode.id_] = nodeParents;
-}
   getGraph = () => {
     let graphName = this.props.graphName.replace("-", " ");
     let url = new URL("/get-json-data", document.location);
@@ -210,8 +144,6 @@ export default class Graph extends React.Component {
         var inEdgesObj = {};
         var childrenObj = {};
         var outEdgesObj = {};
-        var hybridParentsObj = {};
-        var hybridRelationships = [];
         var labelsList = data.texts.filter(function(entry) {
           return entry.rId.startsWith("tspan");
         });
@@ -234,11 +166,6 @@ export default class Graph extends React.Component {
           }
         });
 
-        hybridsList.forEach(hybrid => {
-          childrenObj[hybrid.id_] = [];
-          outEdgesObj[hybrid.id_] = [];
-          this.populateParentsAndHybridRelationship(hybrid, nodesList, hybridRelationships, hybridParentsObj);
-        })
 
         nodesList.forEach(node => {
           parentsObj[node.id_] = [];
@@ -246,12 +173,13 @@ export default class Graph extends React.Component {
           childrenObj[node.id_] = [];
           outEdgesObj[node.id_] = [];
 
-          hybridRelationships.forEach(element => {
-            if (element[0] === node.id_){
-              childrenObj[node.id_].push(element[1]);
-            }
-          });
         });
+
+        hybridsList.forEach(hybrid => {
+          childrenObj[hybrid.id_] = [];
+          outEdgesObj[hybrid.id_] = [];
+          populateHybridRelatives(hybrid, nodesList, parentsObj, childrenObj);
+        })
 
         edgesList.forEach(edge => {
           if (edge.target in parentsObj) {
@@ -284,7 +212,6 @@ export default class Graph extends React.Component {
             'inEdges': inEdgesObj,
             'children': childrenObj,
             'outEdges': outEdgesObj,
-            'hybridParents': hybridParentsObj
           }
         });
       })
@@ -835,6 +762,173 @@ export default class Graph extends React.Component {
       </div>
     );
   }
+}
+
+/** Helper function for parsing hybrid node's text
+ *
+ * @param {string} s
+ * @returns {Array}
+ */
+function parseAnd(s) {
+  "use strict";
+
+  var curr = s;
+  var andList = [];
+  while (curr.length > 0) {
+    if (
+      curr.charAt(0) === "," ||
+      curr.charAt(0) === ";" ||
+      curr.charAt(0) === " "
+    ) {
+      curr = curr.substr(1);
+    } else {
+      var result = parseOr(curr);
+      if (curr === result[1]) {
+        console.error("Parsing failed for " + s + "  with curr = " + curr);
+        break;
+      } else {
+        curr = result[1];
+        andList.push(result[0]);
+      }
+    }
+  }
+  return [andList, curr];
+}
+
+/**
+ * Helper function for parsing hybrid node's text
+ * @param {string} s
+ * @returns {Array}
+ */
+function parseOr(s) {
+  "use strict";
+
+  var curr = s;
+  var orList = [];
+  var tmp;
+  var result;
+  var coursePrefix;
+  while (curr.length > 0 && curr.charAt(0) !== "," && curr.charAt(0) !== ";") {
+    if (curr.charAt(0) === "(") {
+      tmp = curr.substr(1, curr.indexOf(")"));
+      if (coursePrefix === undefined && tmp.length >= 6) {
+        coursePrefix = tmp.substr(0, 3).toUpperCase();
+      }
+      result = parseCourse(tmp, coursePrefix);
+
+      orList.append(result[0]);
+      curr = curr.substr(curr.indexOf(")") + 1);
+    } else if (curr.charAt(0) === " " || curr.charAt(0) === "/") {
+      curr = curr.substr(1);
+    } else {
+      if (coursePrefix === undefined && curr.length >= 6) {
+        coursePrefix = curr.substr(0, 3).toUpperCase();
+      }
+      result = parseCourse(curr, coursePrefix);
+      if (curr === result[1]) {
+        console.error("Parsing failed for " + s + " with curr = " + curr);
+        break;
+      }
+      curr = result[1];
+      orList.push(result[0]);
+    }
+  }
+
+  if (orList.length === 1) {
+    orList = orList[0];
+  }
+
+  return [orList, curr];
+}
+
+/**
+ * Helper function for parsing hybrid node's text
+ * @param {string} s
+ * @param {string} prefix
+ * @returns {Array}
+ */
+function parseCourse(s, prefix) {
+  "use strict";
+
+  var start = s.search(/[,/]/);
+
+  if (start === 3) {
+    return [prefix + s.substr(0, start), s.substr(start)];
+  } else if (start > 0) {
+    return [s.substr(0, start).toUpperCase(), s.substr(start)];
+  }
+
+  if (s.length === 3) {
+    return [prefix + s, ""];
+  }
+
+  return [s, ""];
+}
+
+
+  /** Helper function that adds parents of hybridNode to the parents object, and adds hybrid nodes as children of the Nodes they represent
+ *
+ * @param {Node} hybridNode
+ * @param {Array} nodesJSON
+ * @param {Object} parents
+ * @param {Object} childrenObj
+ */
+function populateHybridRelatives(hybridNode, nodesJSON, parents, childrenObj){
+
+  /**
+   * Helper for hybrid computation. Finds the node with the same course label as the hybrid.
+   * @param  {string} course
+   * @param {Array} nodesJSON
+   * @return {Node}
+   */
+  var findRelationship = (course, nodesJSON) => {
+    var nodes = nodesJSON;
+    var node = nodes.find(
+      n =>
+        n.type_ === "Node" &&
+        n.text.some(textTag => textTag.text.includes(course))
+    );
+    return node;
+  }
+  // parse prereqs based on text
+  var hybridText = "";
+  hybridNode.text.forEach(textTag => (hybridText += textTag.text));
+  var nodeParents = [];
+  // First search for entire string (see Stats graph)
+  var prereqNode = findRelationship(hybridText, nodesJSON);
+  if (prereqNode !== undefined) {
+    nodeParents.push(prereqNode.id_);
+    childrenObj[prereqNode.id_].push(hybridNode.id_);
+  } else {
+    // Parse text first
+    var prereqs = parseAnd(hybridText)[0];
+    prereqs.forEach(course => {
+      if (typeof course === "string") {
+        prereqNode = findRelationship(course, nodesJSON);
+        if (prereqNode !== undefined) {
+          nodeParents.push(prereqNode.id_);
+          childrenObj[prereqNode.id_].push(hybridNode.id_);
+        } else {
+          console.error("Could not find prereq for ", hybridText);
+        }
+      } else if (typeof course === "object") {
+        var orPrereq = [];
+        course.forEach(c => {
+          var prereqNode = findRelationship(c, nodesJSON);
+          if (prereqNode !== undefined) {
+            orPrereq.push(prereqNode.id_);
+            childrenObj[prereqNode.id_].push(hybridNode.id_);
+          } else {
+            console.error("Could not find prereq for ", hybridText);
+          }
+        });
+        if (orPrereq.length > 0) {
+          nodeParents.push(orPrereq);
+        }
+      }
+    });
+  }
+parents[hybridNode.id_] = nodeParents;
 }
 
 Graph.propTypes = {
