+++ js/components/graph/NodeGroup.js
@@ -48,11 +48,9 @@ export default class NodeGroup extends React.Component {
   render() {
     var svg = this.props.svg;
     var highlightedNodes = this.props.highlightedNodes;
-    var hybridRelationships = [];
     return (
       <g id="nodes">
         {this.props.hybridsJSON.map(entry => {
-          var parents = this.props.connections.hybridParents[entry.id_];
           return (
             <Node
               JSON={entry}
@@ -60,7 +58,7 @@ export default class NodeGroup extends React.Component {
               key={entry.id_}
               hybrid={true}
               ref={this.setRefEntry(entry)}
-              parents={parents}
+              parents={this.props.connections.parents[entry.id_]}
               childs={this.props.connections.children[entry.id_]}
               inEdges={[]}
               outEdges={this.props.connections.outEdges[entry.id_]}
@@ -72,12 +70,6 @@ export default class NodeGroup extends React.Component {
         {this.props.nodesJSON.map(entry => {
           var highlighted = highlightedNodes.indexOf(entry.id_) >= 0;
           var childs = this.props.connections.children[entry.id_].slice();
-          hybridRelationships.forEach(element => {
-
-            if (element[0] === entry.id_) {
-              childs.push(element[1]);
-            }
-          });
           return (
             <Node
               JSON={entry}
@@ -104,108 +96,7 @@ export default class NodeGroup extends React.Component {
   }
 }
 
-/**
- *
- * @param {string} s
- * @returns {Array}
- */
-function parseAnd(s) {
-  "use strict";
 
-  var curr = s;
-  var andList = [];
-  while (curr.length > 0) {
-    if (
-      curr.charAt(0) === "," ||
-      curr.charAt(0) === ";" ||
-      curr.charAt(0) === " "
-    ) {
-      curr = curr.substr(1);
-    } else {
-      var result = parseOr(curr);
-      if (curr === result[1]) {
-        console.error("Parsing failed for " + s + "  with curr = " + curr);
-        break;
-      } else {
-        curr = result[1];
-        andList.push(result[0]);
-      }
-    }
-  }
-  return [andList, curr];
-}
-
-/**
- *
- * @param {string} s
- * @returns {Array}
- */
-function parseOr(s) {
-  "use strict";
-
-  var curr = s;
-  var orList = [];
-  var tmp;
-  var result;
-  var coursePrefix;
-  while (curr.length > 0 && curr.charAt(0) !== "," && curr.charAt(0) !== ";") {
-    if (curr.charAt(0) === "(") {
-      tmp = curr.substr(1, curr.indexOf(")"));
-      if (coursePrefix === undefined && tmp.length >= 6) {
-        coursePrefix = tmp.substr(0, 3).toUpperCase();
-      }
-      result = parseCourse(tmp, coursePrefix);
-
-      orList.append(result[0]);
-      curr = curr.substr(curr.indexOf(")") + 1);
-    } else if (curr.charAt(0) === " " || curr.charAt(0) === "/") {
-      curr = curr.substr(1);
-    } else {
-      if (coursePrefix === undefined && curr.length >= 6) {
-        coursePrefix = curr.substr(0, 3).toUpperCase();
-      }
-      result = parseCourse(curr, coursePrefix);
-      if (curr === result[1]) {
-        console.error("Parsing failed for " + s + " with curr = " + curr);
-        break;
-      }
-      curr = result[1];
-      orList.push(result[0]);
-    }
-  }
-
-  if (orList.length === 1) {
-    orList = orList[0];
-  }
-
-  return [orList, curr];
-}
-
-/**
- *
- * @param {string} s
- * @param {string} prefix
- * @returns {Array}
- */
-function parseCourse(s, prefix) {
-  "use strict";
-
-  var start = s.search(/[,/]/);
-
-  if (start === 3) {
-    return [prefix + s.substr(0, start), s.substr(start)];
-  } else if (start > 0) {
-    return [s.substr(0, start).toUpperCase(), s.substr(start)];
-  }
-
-  if (s.length === 3) {
-    return [prefix + s, ""];
-  }
-
-  return [s, ""];
-}
-
-export {parseAnd};
 
 NodeGroup.propTypes = {
   edgesJSON: PropTypes.array,
